authOptions 관련 build 문제 수정
