+++ src/app/api/profile/edit/route.ts
@@ -1,9 +1,9 @@
 import { NextResponse } from "next/server";
 import bcrypt from "bcryptjs";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { ObjectId } from "mongodb";
 import clientPromise from "../../../../../lib/mongodb";
+import { authOptions } from "../../../../../lib/authOptions";
 
 export async function PATCH(req: Request) {
   const session = await getServerSession(authOptions);
